requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
bidict==0.22.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
idna==3.4
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-engineio==4.7.1
python-socketio==5.9.0
PyYAML==6.0.1
redis==5.0.1
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
wsproto==1.2.0
